backend/requirements.txt
Flask==2.0.3
kafka-python==2.0.2
marshmallow==3.14.1
Werkzeug==2.0.3
Flask-CORS==3.0.10
pyspark
numpy
tensorflow
scikit-learn
